chore: Specify target version for ruff
Import `annotations` from the `__future__` module to enable the package
to work seamlessly on older versions of python.  Also specify the target
version a 3.9, such that when we turn on pyupgrade linting/formatting,
we can bump that target version and ruff will take care of the changes
for us.

Also enable more ruff linters.