refactor: Idiomatic pathing
When OpenCSP was created, the paradigm was to use separate strings to
represent the directory containing a file, the file name (without an
extension), and the extension.  However, these different path components
can be handled natively with `pathlib.Path` objects.  The prior commit
switched over from `os.path` operations to `Path` objects, but there
were still remnants of the older paradigm lying around.  This commit
cleans that up so any conceptual file or directory is represented by a
`Path` object, and if you need to get at the directory, name, or
extension, you can do so by accessing its properties.

This commit also contains the following refactorings:

* Moves the `from_csv()` method into the `CsvInterface` base class.
* Reverses `if` conditions to dedent content.