refactor: Pull method out to module function
The `to_csv` method felt odd, in that it needed to be called on a
`CsvInterface` object, but then passed a list of objects of the same
type.  It feels cleaner to have this be a module function that is just
passed the list of objects.  There's a new check to ensure all the
objects are of the same type, so you know the header you're getting from
the first object corresponds to the values you're getting for all the
objects (e.g., you wouldn't want to write a `FileCache` header and then
a `FileFingerprint` data line).