refactor: Extract configuration
Pull the argument parser and associated methods and attributes out into
a new `Config` class.  This better delineates between attributes that
are user-supplied configuration values, which don't change once set, vs
those that are used to modify data as its passed between methods of the
`SensitiveStringsSearcher` class.

This commit also removes the `clone()` methods, because it would have
been awkward to modify it to work with this new `config` attribute.  I
still don't like how we currently create a new searcher for each HDF5
file, so that'll likely be refactored in the future, but leaving
`clone()` in place would've been worse than inlining it as was done
here.