refactor: Extract user interaction methods
Pull them out into a separate class.  I'm doing this so it'll be
possible to extract the image search, binary search, hdf5 search, etc.,
out into separate files.

Note that I don't like having to couple the `assume_yes` in this class
to the one in `Config`.  I'd thought about creating a `config` global
object to be shared across all files, but the creation of a separate
`SensitiveStringsSearcher` with its own unique `config` to search each
HDF5 file means this solution isn't available to us.  There's likely a
better way to do this; I just haven't thought of it yet.